Refactors custom vocabulary handling
Consolidates custom vocabulary logic into a single `CustomVocabulary` utility.

This change removes duplicate vocabulary retrieval code from individual transcription services and centralizes it in the `CustomVocabulary` enum, promoting code reuse and maintainability. The Deepgram service is also updated to respect the maximum number of keywords recommended by the service.
